Report test + asset quantity as double

+++ src/main/java/crypto_wallet/domain/data/CryptoAsset.java
@@ -2,4 +2,4 @@ package crypto_wallet.domain.data;
 
 import java.math.BigDecimal;
 
-public record CryptoAsset(String symbol, int quantity, BigDecimal price) {}
+public record CryptoAsset(String symbol, double quantity, BigDecimal price) {}
